+++ snap/blocks.js
@@ -462,7 +462,7 @@ export class IconView {
     this.height = 12
     this.modified = false
     this.name = ""
-    this.color = new Color(255, 255, 255)
+    this.color = null
     this.padx = 4
     this.fillAttribute = "fill"
 
@@ -480,7 +480,7 @@ export class IconView {
     }
 
     if (!this.color) {
-      this.color = info.color || new Color(255, 255, 255)
+      this.color = info.color || null
     }
 
     if (!this.fillAttribute) {
@@ -553,7 +553,10 @@ export class IconView {
    * @param {Options} options
    * @returns {SVGElement}
    */
-  draw(options) {
+  draw(options, parent) {
+    if (!this.color) {
+      this.color = parent.isZebra ? new Color() : new Color(255,255,255)
+    }
     let props = {
       width: this.width,
       height: this.height,