+++ HISTORY.md
@@ -5,6 +5,7 @@
 - [x] Block prototypes in Japanese translation
 - [x] Block highlighting
 - [x] Fix viewport height on single scratch 3 define block (pulled from scratchblocks).
+- [x] Default icon color is not affected by zebra coloring
 
 # 1.8.1
 ## Additions